pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "generates diff of two files"
authors = ["maryker mashalipatowa@yandex.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.8.1"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"


[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
